Moving mutator into shared module.
