+++ eventsummary/build.gradle.kts
@@ -20,6 +20,9 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 api(project(":core-ui"))
+                implementation(project(":event-api"))
+                implementation("com.tunjid.mutator:core:${Versions.mutator}")
+                implementation("com.tunjid.mutator:coroutines:${Versions.mutator}")
             }
         }
         val commonTest by getting {
